fix(#984): exposes configuration to initialize K8Client in Kuberenetes/Openshift Requirement.
