+++ kubernetes/kubernetes/pom.xml
@@ -114,6 +114,10 @@
       <groupId>org.awaitility</groupId>
       <artifactId>awaitility</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.jboss.arquillian.config</groupId>
+      <artifactId>arquillian-config-impl-base</artifactId>
+    </dependency>
 
   </dependencies>
