+++ kubernetes/kubernetes/src/main/java/org/arquillian/cube/kubernetes/impl/ClientCreator.java
@@ -15,16 +15,16 @@
  */
 package org.arquillian.cube.kubernetes.impl;
 
-import io.fabric8.kubernetes.api.builder.v3_1.TypedVisitor;
 import io.fabric8.kubernetes.clnt.v3_1.ConfigBuilder;
 import io.fabric8.kubernetes.clnt.v3_1.DefaultKubernetesClient;
 import io.fabric8.kubernetes.clnt.v3_1.KubernetesClient;
-import org.arquillian.cube.impl.util.Strings;
 import org.arquillian.cube.kubernetes.api.Configuration;
 import org.jboss.arquillian.core.api.InstanceProducer;
 import org.jboss.arquillian.core.api.annotation.ApplicationScoped;
 import org.jboss.arquillian.core.api.annotation.Inject;
 import org.jboss.arquillian.core.api.annotation.Observes;
+
+import static org.arquillian.cube.kubernetes.ClientConfigurator.getConfigBuilder;
 
 /**
  * Creates an instances of the {@link KubernetesClient} when a {@link @Configuration} is available.
@@ -36,31 +36,7 @@ public class ClientCreator {
     private InstanceProducer<KubernetesClient> producer;
 
     public void createClient(@Observes Configuration config) {
-
-        final ConfigBuilder configBuilder = new ConfigBuilder()
-            .withNamespace(config.getNamespace())
-            .withApiVersion(config.getApiVersion())
-            .withTrustCerts(config.isTrustCerts())
-            .accept(new TypedVisitor<ConfigBuilder>() {
-                @Override
-                public void visit(ConfigBuilder b) {
-                    b.withNoProxy(b.getNoProxy() == null ? new String[0] : b.getNoProxy());
-                }
-            });
-
-        if (Strings.isNotNullOrEmpty(config.getMasterUrl().toString())) {
-            configBuilder.withMasterUrl(config.getMasterUrl().toString());
-        }
-
-        if (Strings.isNotNullOrEmpty(config.getToken())) {
-            configBuilder.withOauthToken(config.getToken());
-        }
-
-        if (Strings.isNotNullOrEmpty(config.getUsername()) && Strings.isNotNullOrEmpty(config.getPassword())) {
-            configBuilder.withUsername(config.getUsername())
-                .withPassword(config.getPassword());
-        }
-
+        final ConfigBuilder configBuilder = getConfigBuilder(config);
         producer.set(new DefaultKubernetesClient(configBuilder.build()));
     }
 }
