+++ pom.xml
@@ -67,7 +67,7 @@
   </issueManagement>
 
   <properties>
-    <version.arquillian_core>1.2.0.Final</version.arquillian_core>
+    <version.arquillian_core>1.4.0-SNAPSHOT</version.arquillian_core>
     <version.junit>4.12</version.junit>
     <version.hamcrest>1.3</version.hamcrest>
     <version.docker-java>3.0.14</version.docker-java>
